demo files + minor fixes

+++ src/groups/z_bmq/z_bmqa/z_bmqa_message.h
@@ -5,7 +5,7 @@ typedef struct z_bmqa_Message z_bmqa_Message;
 
 int z_bmqa_Message__createEmpty(z_bmqa_Message** message_obj);
 
-int z_bmqa_Message_setDataRef(z_bmqa_Message** message_obj, const char* data, size_t length);
+int z_bmqa_Message_setDataRef(z_bmqa_Message** message_obj, const char* data, int length);
 
 
 #endif
